backend/requirements.txt
uvicorn[standard]==0.27.0
fastapi==0.109.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
networkx==3.2.1
numpy==1.26.3
python-multipart==0.0.6
geopy==2.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
